+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # 2.3.5
-1. 增加token命令，创建Qbox token, qiniuToken, uploadToken and DownloadToken
+1. 增加token命令，创建Qbox token, qiniuToken, uploadToken
 2. qshell配置文件支持设置accessKey和secretKey
 3. 增加了arm版本
 4. listbucket2增加文件大小的可读性，可以使用Bytes, KB, MB, GB格式显示